build: cast implicit narrowing conversions and reorder BLEManager init list
- Add explicit static_cast<uint8_t>/<uint16_t> around arithmetic that
  narrows into packed struct fields and ring-buffer indices, silencing
  -Wconversion warnings in types.h and motor_event_buffer.cpp
- Fix EOT_CHAR to use a char literal instead of an int literal
- Reorder BLEManager member-initializer list to match declaration order
- Drop compound increment on volatile-adjacent overflow counters in
  sync_protocol.cpp in favor of explicit assignment
- Cast patternType to int for Serial.printf's %d in therapy_engine.cpp

+++ include/types.h
@@ -456,14 +456,14 @@ struct __attribute__((packed)) MacrocycleEvent {
         , finger(secFinger)
         , amplitude(amp)
         , durationMs(dur)
-        , freqOffset((freqHz - MACROCYCLE_FREQ_BASE) / MACROCYCLE_FREQ_STEP)
+        , freqOffset(static_cast<uint8_t>((freqHz - MACROCYCLE_FREQ_BASE) / MACROCYCLE_FREQ_STEP))
         , primaryFinger(primFinger) {}
 
     /**
      * @brief Decode frequency from freqOffset
      */
     uint16_t getFrequencyHz() const {
-        return MACROCYCLE_FREQ_BASE + (freqOffset * MACROCYCLE_FREQ_STEP);
+        return static_cast<uint16_t>(MACROCYCLE_FREQ_BASE + (freqOffset * MACROCYCLE_FREQ_STEP));
     }
 };
